+++ src/render/mod.rs
@@ -17,11 +17,7 @@ pub(crate) use material::TerminalMaterialPlugin;
 pub(crate) use mesh::TerminalMeshPlugin;
 pub(crate) use uv_mapping::TerminalUvMappingPlugin;
 
-pub use camera::{
-    //TerminalSystemsCacheCameraData,
-    TerminalSystemsUpdateCamera,
-    UpdateTerminalViewportEvent,
-};
+pub use camera::{TerminalSystemsUpdateCamera, UpdateTerminalViewportEvent};
 pub use font::TerminalSystemsUpdateFont;
 pub use mesh::TerminalSystemsUpdateMesh;